Update nan values method
In order to compact the code and the generated files, the virtual MAC presence is configured to add this virtual MAC with 0 messages in case of the MAC is not found.

Additionally, for each interval that any Raspberry was disabled, a nan row is added for this Raspberry at this interval, so it facilitates the nan values treatment.